Added distributed params and optimizer. Still adapting for better torch usability and huggingface support.

+++ tensorlink/mpc/nodes.py
@@ -1,12 +1,14 @@
 from tensorlink.roles.user import User
 from tensorlink.roles.validator import Validator
 from tensorlink.roles.worker import Worker
-from tensorlink.ml.distributed import DistributedModel
+from tensorlink.ml.module import DistributedModel
 from tensorlink.ml.worker import DistributedWorker
+from tensorlink.ml.optim import create_distributed_optimizer
+from tensorlink.ml.graphing import handle_layers
+from tensorlink.ml.utils import access_module
 
 import multiprocessing
-import threading
-import atexit
+import torch
 import time
 
 # Set the start method to 'spawn'
@@ -37,16 +39,25 @@ class BaseNode:
 
         self._initialized = True
         self.setup()
-        atexit.register(self.cleanup)
 
     def setup(self):
-        self.node_process = multiprocessing.Process(target=self.run_role)
+        self.node_process = multiprocessing.Process(target=self.run_role, daemon=True)
         self.node_process.start()
 
     def cleanup(self):
         if self.node_process is not None:
-            self.node_process.terminate()
+            # Send a stop request to the role instance
+            self.send_request("stop", (None,))
+            self.node_process.join(timeout=10)
+
+            # If the process is still alive, terminate it
+            if self.node_process.is_alive():
+                print("Forcing termination for DistributedWorker process.")
+                self.node_process.terminate()
+
+            # Final join to ensure it's completely shut down
             self.node_process.join()
+            self.node_process = None  # Reset to None after cleanup
 
     def send_request(self, request_type, args):
         """
@@ -92,7 +103,7 @@ class UserNode(BaseNode):
     def run_role(self):
         kwargs = self.init_kwargs.copy()
         kwargs.update({
-            'debug': kwargs.get('debug', True),
+            'debug': kwargs.get('debug', False),
             'upnp': kwargs.get('upnp', True),
             'off_chain_test': kwargs.get('off_chain_test', False)
         })
@@ -104,24 +115,40 @@ class UserNode(BaseNode):
         role_instance.start()
         self.node_process = role_instance
 
-    def create_distributed_model(self, model, n_pipelines, dp_factor, config=None):  # TODO Max module size etc?
-        dist_model = DistributedModel(self.node_requests, self.node_responses, self.mpc_lock, model, dp_factor)
-
+    def create_distributed_model(self, model, n_pipelines, optimizer_type=None, dp_factor=None):  # TODO Max module size etc?
+        dist_model = DistributedModel(self.node_requests, self.node_responses, self.mpc_lock, model, n_pipelines)
         self.send_request("request_workers", None)
         time.sleep(3)
         workers = self.send_request("check_workers", None)
+        if len(workers) == 0:
+            time.sleep(5)
+            workers = self.send_request("check_workers", None)
+
+        if optimizer_type is None:
+            optimizer_type = torch.optim.Adam
+
         dist_model.worker_info = workers
-        distribution = dist_model.parse_model(model, handle_layer=True)
-        distributed_config = self.send_request("request_job", (n_pipelines, dp_factor, distribution))
+        distribution = dist_model.parse_model(model, handle_layer=False)
+        distributed_config = self.send_request("request_job", (n_pipelines, 1, distribution))
+
+        for module_id, module_info in distributed_config.items():
+            if module_info["type"] == "offloaded":
+                module, module_name = access_module(model, module_info["mod_id"])
+                setattr(module, "optimizer", optimizer_type)
+
         dist_model.distribute_model(distributed_config)
-        return dist_model
+
+        def _create_distributed_optimizer(**optimizer_kwargs):
+            return create_distributed_optimizer(dist_model, optimizer_type, **optimizer_kwargs)
+
+        return dist_model, _create_distributed_optimizer
 
 
 class WorkerNode(BaseNode):
     def run_role(self):
         kwargs = self.init_kwargs.copy()
         kwargs.update({
-            'debug': kwargs.get('debug', True),
+            'debug': kwargs.get('debug', False),
             'upnp': kwargs.get('upnp', True),
             'off_chain_test': kwargs.get('off_chain_test', False)
         })
